emacs: Add wrapped emacs for epg issues
This is just a hack to avoid https://dev.gnupg.org/T6481. What happens
is that the order of output from gpg has changed between gnupg `2.4.0`
and `2.4.1`. This causes emacs to hang as it's unable to detect that the
gpg process is ready for input.

This hack only provides a separate executable `epg` which sets `$PATH`
to point to a `2.4.0` version of `gnupg`. Which means that to have
working encryption, emacs must be launched with `epg`.

I don't want to miss out when this suddenly starts working again, so
it's better to make this an "opt-in" thing.