emacs: Copilot chose to be difficult again...
So there I was, happily minding my own business, upgrading some
packages, when all of a sudden ... Emacs won't start.

Or, `emacs` runs, but starting an `emacs.service` server in `systemd`
for `emacsclient` to connect to wouldn't start. And in regular `emacs`
things like highlighting and whatever didn't work properly.

Recently this has been due to issues with the
[copilot.el](https://github.com/copilot-emacs/copilot.el) package, and
in particular on NixOS.

OK, as with so much other software, `copilot.el` makes quite a lot of
assumptions that aren't really compatible with NixOS. I don't *have*
`nodejs` and `npm` in `$PATH` ... for a *reason*.

Long story short, just wrap it with `nodejs` and set
`copilot-server-executable` explicitly.